Model creation and Prediction Updates
made changes to the model builder and model predictor. Now the model
builder should be able to create two models and save that to the
database and the predictor should be able to predict late delivery based
on shipment input. i also edited the model metadata. java so it saves
the correct model information to the database. Included getters and
setters for the shipment.java